Unsubscribe from document updates when a user is removed from a document
